+++ src/cmdline/install.rs
@@ -272,7 +272,7 @@ impl InstallConfig {
             return Ok(self);
         }
 
-        let args = self
+        let mut args = self
             .config_file
             .iter()
             .map(|path| {
@@ -294,6 +294,12 @@ impl InstallConfig {
                     .context("serializing command-line arguments")?,
             )
             .collect::<Vec<_>>();
+
+        // If firstboot-args is defined, add it manually
+        if let Some(firstboot_args) = &self.firstboot_args {
+            args.push("--firstboot-args".to_string());
+            args.push(firstboot_args.clone());
+        }
 
         println!("Running with arguments: {}", args.join(" "));
         Self::from_args(&args)
@@ -577,5 +583,31 @@ dest-device: u
         .unwrap()
         .expand_config_files()
         .unwrap_err();
+    }
+
+    /// Test that firstboot-args is manually added to args list when defined
+    #[test]
+    fn test_firstboot_args_manually_added() {
+        let mut f = NamedTempFile::new().unwrap();
+        f.as_file_mut().write_all(b"dest-device: /dev/sda").unwrap();
+
+        let config = InstallConfig::from_args(&[
+            "--config-file",
+            f.path().to_str().unwrap(),
+            "--firstboot-args",
+            "ip=dhcp",
+        ])
+        .unwrap();
+
+        // Verify firstboot-args is defined
+        assert!(config.firstboot_args.is_some());
+        assert_eq!(config.firstboot_args.as_ref().unwrap(), "ip=dhcp");
+
+        // Test expand_config_files to verify manual addition
+        let expanded = config.expand_config_files().unwrap();
+
+        // Should still have firstboot-args
+        assert!(expanded.firstboot_args.is_some());
+        assert_eq!(expanded.firstboot_args.unwrap(), "ip=dhcp");
     }
 }